# openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "Payment-Signature",
        "description": "x402 v2 payment payload (base64)"
      }
    }
  },
  "info": {
    "title": "x402 Trust Layer",
    "version": "5.2.1",
    "description": "68 paid x402 trust infrastructure APIs: guard, semantic escrow, mandate diff, certified seller network, KYM, mandates, compliance, disputes, and orchestration.",
...
  },
...
  "servers": [
    {
      "url": "https://x402trustlayer.xyz"
    }
  ],
  "paths": {
...
    "/api/refund-arbiter/evaluate": {
      "post": {
        "operationId": "post_refund_arbiter_evaluate",
        "summary": "Evaluate refund eligibility from verification signals",
        "description": "Evaluate refund eligibility from verification signals \u2014 $0.08 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "trust",
          "refund",
          "arbiter"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "allowed": true,
                  "summary": "Paid response after x402 settlement"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.080000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "verificationScore": {
                    "type": "number"
                  },
                  "responseEmpty": {
                    "type": "boolean"
                  },
                  "responseGeneric": {
                    "type": "boolean"
                  },
                  "endpointReachable": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "verificationScore",
                  "responseEmpty",
                  "responseGeneric",
                  "endpointReachable"
                ]
              },
              "example": {
                "verificationScore": 85,
                "responseEmpty": false,
                "responseGeneric": false,
                "endpointReachable": true
              }
            }
          }
        }
      }
    },
...
    "/api/compliance/ledger": {
      "post": {
        "operationId": "post_compliance_ledger",
        "summary": "CFO/SOC2-grade spend reconciliation with policy flags and tamper hash",
        "description": "CFO/SOC2-grade spend reconciliation with policy flags and tamper hash \u2014 $0.12 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "tier1",
          "compliance",
          "audit",
          "cfo",
          "ledger"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "allowed": true,
                  "summary": "Paid response after x402 settlement"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.120000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "organizationId": {
                    "type": "string"
                  },
                  "period": {
                    "type": "string"
                  },
                  "records": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "merchant": {
                          "type": "string"
                        },
                        "amountUsdc": {
                          "type": "number"
                        },
                        "rail": {
                          "type": "string"
                        },
                        "network": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string"
                        },
                        "agentId": {
                          "type": "string"
                        },
                        "transactionHash": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "merchant",
                        "amountUsdc",
                        "rail",
                        "network",
                        "category",
                        "agentId",
                        "transactionHash",
                        "timestamp"
                      ]
                    }
                  },
                  "policy": {
                    "type": "object",
                    "properties": {
                      "monthlyCapUsdc": {
                        "type": "number"
                      },
                      "perMerchantCapUsdc": {
                        "type": "number"
                      },
                      "requireTxHash": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "monthlyCapUsdc",
                      "perMerchantCapUsdc",
                      "requireTxHash"
                    ]
                  }
                },
                "required": [
                  "organizationId",
                  "period",
                  "records",
                  "policy"
                ]
              },
              "example": {
                "organizationId": "verifier-org",
                "period": "2026-05",
                "records": [
                  {
                    "merchant": "api.myceliasignal.com",
                    "amountUsdc": 0.05,
                    "rail": "base-x402",
                    "network": "eip155:8453",
                    "category": "oracle",
                    "agentId": "dexter-verifier-probe",
                    "transactionHash": "0x0000000000000000000000000000000000000000000000000000000000000001",
                    "timestamp": "2026-06-08T15:22:47.602Z"
                  }
                ],
                "policy": {
                  "monthlyCapUsdc": 1000,
                  "perMerchantCapUsdc": 500,
                  "requireTxHash": true
                }
              }
            }
          }
        }
      }
    },
    "/api/dispute/resolve": {
      "post": {
        "operationId": "post_dispute_resolve",
        "summary": "Visa chargeback dossier or on-chain refund claim builder",
        "description": "Visa chargeback dossier or on-chain refund claim builder \u2014 $0.10 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "tier1",
          "dispute",
          "chargeback",
          "visa",
          "refund"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "allowed": true,
                  "summary": "Paid response after x402 settlement"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.100000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rail": {
                    "type": "string"
                  },
                  "merchant": {
                    "type": "string"
                  },
                  "amountUsdc": {
                    "type": "number"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "evidence": {
                    "type": "object",
                    "properties": {
                      "actualResponseEmpty": {
                        "type": "boolean"
                      },
                      "receiptValid": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "actualResponseEmpty",
                      "receiptValid"
                    ]
                  }
                },
                "required": [
                  "rail",
                  "merchant",
                  "amountUsdc",
                  "reason",
                  "evidence"
                ]
              },
              "example": {
                "rail": "visa-cli",
                "merchant": "api.myceliasignal.com",
                "amountUsdc": 1,
                "reason": "non_delivery",
                "evidence": {
                  "actualResponseEmpty": true,
                  "receiptValid": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}